pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "grrhs"
version = "0.1.0"
description = "Generalized regularized horseshoe experimentation toolkit"
readme = "README.md"
requires-python = ">=3.9"
license = { file = "LICENSE" }
authors = [
  { name = "GRRHS Team" }
]
dependencies = [
  "numpy>=1.24",
  "scipy>=1.10",
  "jax>=0.4",
  "jaxlib>=0.4",
  "numpyro>=0.13",
  "torch>=2.1",
  "matplotlib>=3.7",
  "pandas>=1.5",
  "PyYAML>=6.0",
  "skglm>=0.5",
  "scikit-learn>=1.3",
  "tqdm>=4.65",
  "polyagamma>=1.3"
]
keywords = ["bayesian", "horseshoe", "variational", "mcmc"]

[project.optional-dependencies]
dev = [
  "pytest>=7.4",
  "pre-commit>=3.6",
  "ruff>=0.5"
]

[tool.setuptools.packages.find]
# Discover packages from repository root so top-level 'grrhs' is included
where = ["."]
